+++ test/libsolidity/smtCheckerTests/operators/bitwise_xor_uint.sol
@@ -1,0 +1,17 @@
+pragma experimental SMTChecker;
+
+contract C {
+	function f() public pure {
+		uint8 x = 1;
+		uint16 y = 0;
+		assert(x ^ y != 0);
+		x = 0xff;
+		y = 0xff00;
+		assert(x ^ y == 0xff);
+		assert(x ^ y == 0xffff);
+		assert(x ^ y == 0x0000);
+	}
+}
+// ----
+// Warning 6328: (155-176): Assertion violation happens here
+// Warning 6328: (207-230): Assertion violation happens here